Skaffold_Deployments/config_log_formatter/requirements.txt
flask
pandas
matplotlib
chardet
pyyaml
